+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/EurecaMapKey.java
@@ -1,4 +1,4 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class EurecaMapKey {
+public interface EurecaMapKey {
 }
