+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/EurecaMapValue.java
@@ -1,4 +1,4 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class EurecaMapValue {
+public interface EurecaMapValue {
 }
